optimize range read performance in multi-node (#543)
1. Use larger scan batch size to decrease number of remote request needed to finish the scan.
2. Balance SetFinished workload of remote scan slice request. SetFinish will be putting up all the scan tuples into a single big byte array, which usually causes around 5-10 microseconds.
3. Use a separate rpc stream for scan slice response so that it would block other fast cc requests.
4. Use tx processor to speed up deserialization if there are too many scan slice response piling up.
5. Flatten scan slice response message by decreasing usage of repeated pointer fields to speed up deserialization.
6. Revert usage of arena since it would cause performance regression on smaller size remote message deserialization.
7. Add a unique_ptr to TemplateScanTuple for storing deserialized tuple from remote scan slice response so that it won't cause new memory allocation  on every scan slice response deserialization.

+++ src/cc/local_cc_shards.cpp
@@ -1190,7 +1190,7 @@ void LocalCcShards::Terminate()
     }
     task_worker_cv_.notify_all();
 
-    for (size_t idx = 0; idx < data_sync_worker_num_; ++idx)
+    for (int idx = 0; idx < data_sync_worker_num_; ++idx)
     {
         data_sync_worker_thds_.at(idx).join();
     }
@@ -1203,7 +1203,7 @@ void LocalCcShards::Terminate()
     }
     flush_worker_cv_.notify_all();
 
-    for (size_t idx = 0; idx < flush_worker_num_; ++idx)
+    for (int idx = 0; idx < flush_worker_num_; ++idx)
     {
         flush_worker_thds_.at(idx).join();
     }
